% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ridge.R
\name{ridge}
\alias{ridge}
\title{Fit a linear model using ridge regularization}
\usage{
ridge(X, y, lambda_vals)
}
\arguments{
\item{X}{A data frame containing the variables in the model. The levels of categorical variables should be
represented as numeric.}

\item{y}{A vector of the quantitative response variable.}

\item{lambda_vals}{A vector of possible values of the tuning paramenter \eqn{\lambda} in ridge regression.}
}
\value{
\code{ridge} returns a matrix of regression coefficients with ncol(X) columns
  and length(lambda_vals) rows.
}
\description{
\code{ridge} returns the coefficients of the linear models fiited with different values of penalized parameter \eqn{\lambda}.
}
\examples{
data(iris)
X <- iris[,2:5]
X$Species <- as.numeric(X$Species)
y <- as.numeric(iris[,1])
fit <- ridge(X, y, 0:500)
}
\keyword{regression}
\keyword{ridge}
